Add types to files in tests folder
