+++ package.json
@@ -40,6 +40,8 @@
   ],
   "license": "MIT",
   "devDependencies": {
+    "@types/chai": "^4.2.14",
+    "@types/chai-http": "^4.2.0",
     "@types/events": "^3.0.0",
     "@types/express": "^4.17.9",
     "@types/mocha": "^8.0.4",
