+++ tests/1_unit-tests.ts
@@ -6,8 +6,8 @@
  *       (if additional are added, keep them at the very end!)
  */
 
-var chai = require("chai");
+import * as chai from "chai";
 
-suite("Unit Tests", function () {
+suite("Unit Tests", function (): void {
   //none required
 });
