+++ tsconfig.test.json
@@ -4,5 +4,5 @@
     "allowJs": true,
     "target": "es5"
   },
-  "include": ["./*"]
+  "include": ["./**/*"]
 }
